TASK: Cleanup CK editor implementation structure and debounce toolbar updates
