+++ Resources/Private/JavaScript/Guest/Components/Editors/CKEditor/Selection/index.js
